+++ src/forze_postgres/__init__.py
@@ -6,6 +6,7 @@ require_psycopg()
 
 # ....................... #
 
+from ._logging import FORZE_POSTGRES_LOGGER_NAMES, ForzePostgresLogger
 from .adapters import PostgresSchemaTenantProvisioner
 from .execution import (
     POSTGRES_CLIENT_CAPABILITY,
@@ -49,6 +50,8 @@ from .kernel.relation import (
 # ----------------------- #
 
 __all__ = [
+    "FORZE_POSTGRES_LOGGER_NAMES",
+    "ForzePostgresLogger",
     "PostgresDepsModule",
     "PostgresSchemaTenantProvisioner",
     "PostgresClient",